@@ -30,7 +30,7 @@
     "watch:test": "npm run test -- --watch"
   },
   "dependencies": {
-    "abstracted-firebase": "^0.6.3",
+    "abstracted-firebase": "^0.11.8",
     "common-types": "^1.7.36",
     "firebase-admin": "^5.12.1",
     "firebase-api-surface": "^0.8.3",
@@ -43,7 +43,7 @@
   "peerDependencies": {
     "chance": "^1.0.13",
     "faker": "^4.1.0",
-    "firemock": "^0.17.3"
+    "firemock": "^0.17.5"
   },
   "devDependencies": {
     "@types/chai": "^4.1.2",
@@ -61,16 +61,16 @@
     "chance": "^1.0.13",
     "coveralls": "^2.0.0",
     "faker": "^4.1.0",
-    "firemock": "^0.17.3",
+    "firemock": "^0.17.5",
     "js-yaml": "^3.10.0",
     "mocha": "^5.0.5",
     "nyc": "^11.6.0",
     "rimraf": "^2.6.2",
     "rollup": "^0.60.1",
     "test-console": "^1.0.0",
-    "ts-node": "^4.1.0",
+    "ts-node": "^6.1.0",
     "tslint": "^5.10.0",
-    "typescript": "^2.9.1"
+    "typescript": "^3.0.0-dev.20180609"
   },
   "engines": {
     "node": ">=4.0.0"
