@@ -1,6 +1,5 @@
-import { DB } from "./db";
 export { SerializedQuery } from "serialized-query";
-export { DB, IFirebaseConfig, IFirebaseListener, Snapshot, Query, Reference } from "./db";
+export { DB, IFirebaseListener, Snapshot, Query, Reference } from "./db";
 
-import { RealTimeDB, FirebaseBoolean } from "abstracted-firebase";
-export { RealTimeDB, FirebaseBoolean };
+// import { RealTimeDB, FirebaseBoolean, IFirebaseConfig } ;
+export { RealTimeDB, FirebaseBoolean, IFirebaseConfig } from "abstracted-firebase";
